Added user profile page and tweaked tables page
User profile page added. Each user now can see a history of their
searches, with links to view those searches again. To do this, I added a
UserHistory model class in models.py. Django takes care of storing this
model in the backend.

On the tables page, the search box will now come populated with the
value that was last searched.